+++ bp-app-api/src/main/java/com/beautify_project/bp_app_api/config/JasyptConfigBean.java
@@ -1,4 +1,4 @@
-package com.beautify_project.bp_app_api.bean;
+package com.beautify_project.bp_app_api.config;
 
 import lombok.RequiredArgsConstructor;
 import org.jasypt.encryption.StringEncryptor;
